requirements_production.txt
# Requirements para producción
gradio==4.44.0
openai==1.51.0
python-dotenv==1.0.0
langchain==0.3.0
langchain-openai==0.2.0
langchain-community==0.3.0
faiss-cpu==1.8.0
pydantic==2.5.0
requests==2.31.0
sqlite3  # Built-in Python
numpy==1.24.3
pandas==2.0.3
PyYAML==6.0.1